Move Smoke Test and Cli test to semantic package
