+++ tests/integration/semantics/src/test/java/org/apache/pulsar/tests/integration/smoke/SmokeTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.tests.integration;
+package org.apache.pulsar.tests.integration.smoke;
 
 import lombok.Cleanup;
 import org.apache.pulsar.client.api.Consumer;
